Further tidied CMake files as per Han Zhu's requests
Managed to resolve the CMAKE issue around finding OpenSSL, so now
it uses the same method as Casablanca, and doesn't require homebrew
users to specify OPENSSL_ROOT_DIR. The readme explains what to set it
to for non-homebrew users (eg Fink, or Macports)

Tidied up the process for including the -L for gettext, which is
required by libintl, which is part of the dependency chain for
libxml++, but isn't handled very well by pkg-config for some reason.
Again we assume the path for homebrew, (raising messages if we can't)
find it, and the readme describes how non-homebrew users can point
to their installation if needed.

in src/CMakeLists.txt, put an APPLE conditional around adding the warnings
variable to the CXX flags

Tidided up the readme section for OSX so that it drops some of the
linux bits that were copy pasted. Made clearer instructions for homebrew
users vs non-homebrew users. Removed some of the brew packages that the
user needs to install as they will be installed as dependencies if needed.